# docs/static/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "SOARCA API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/manual/": {
            "get": {
                "description": "get all pending manual commands that still needs values to be returned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manual"
                ],
                "summary": "get all pending manual commands that still needs values to be returned",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.InteractionCommandData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.InteractionCommandData"
                            }
                        }
                    }
                }
            }
        },
        "/manual/{exec_id}/{step_execution_id}": {
            "get": {
                "description": "get a specific manual command that still needs a value to be returned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manual"
                ],
                "summary": "get a specific manual command that still needs a value to be returned",
                "parameters": [
                    {
                        "type": "string",
                        "description": "execution ID",
                        "name": "exec_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "step execution ID (identifies a specific pending step invocation; see GET /manual/ to discover it, as multiple pending commands may share the same step ID)",
                        "name": "step_execution_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.InteractionCommandData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "resolve a specific pending manual command by supplying its out args. This is a PUT\non the same resource GET /manual/{exec_id}/{step_execution_id} identifies, not a\ngeneric RPC-style action, so the ids live in the path, not the body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manual"
                ],
                "summary": "resolve a specific pending manual command by supplying its out args",
                "parameters": [
                    {
                        "type": "string",
                        "description": "execution ID",
                        "name": "exec_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "step execution ID (identifies a specific pending step invocation; see GET /manual/ to discover it, as multiple pending commands may share the same step ID)",
                        "name": "step_execution_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "resolution",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ManualOutArgsUpdatePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Execution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "required": [
                "message",
                "original-call",
                "status"
            ],
            "properties": {
                "downstream-call": {
                    "type": "string",
                    "example": "{\"some\" : \"json\"}"
                },
                "message": {
                    "type": "string",
                    "example": "missing argument in call"
                },
                "original-call": {
                    "type": "string",
                    "example": "/example/route"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "api.Execution": {
            "type": "object",
            "required": [
                "execution_id",
                "payload"
            ],
            "properties": {
                "execution_id": {
                    "type": "string",
                    "example": "2c855cd6-bbce-402f-a143-3d6eec346c08"
                },
                "payload": {
                    "type": "string",
                    "example": "playbook--0cec398c-db69-4f17-bde4-8ecbcc4a8879"
                }
            }
        },
        "api.InteractionCommandData": {
            "type": "object",
            "required": [
                "commands",
                "execution_id",
                "out_args",
                "playbook_id",
                "step_execution_id",
                "step_id",
                "targets",
                "type"
            ],
            "properties": {
                "commands": {
                    "description": "All commands of the step, in order. A manual step is a single unit of work resolved by one response, but may list multiple commands/instructions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ManualCommand"
                    }
                },
                "execution_id": {
                    "description": "The id of the execution",
                    "type": "string"
                },
                "out_args": {
                    "description": "Map of cacao variables handled in the step out args with current values and definitions",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.Variables"
                        }
                    ]
                },
                "playbook_id": {
                    "description": "The id of the CACAO playbook executed by the execution",
                    "type": "string"
                },
                "step_execution_id": {
                    "description": "The id of this specific step invocation. Distinguishes concurrent/repeated pending commands that share the same StepId (e.g. overlapping loop iterations)",
                    "type": "string"
                },
                "step_id": {
                    "description": "The id of the step executed by the execution",
                    "type": "string"
                },
                "targets": {
                    "description": "All targets of the step, in order, together with their resolved authentication information (needed by a human operator to perform the step manually)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/capability.ResolvedTarget"
                    }
                },
                "type": {
                    "description": "The type of this content",
                    "type": "string",
                    "example": "execution-status"
                }
            }
        },
        "api.ManualCommand": {
            "type": "object",
            "required": [
                "command",
                "description"
            ],
            "properties": {
                "command": {
                    "description": "The command for the agent, either plain or base64",
                    "type": "string"
                },
                "commandb64": {
                    "description": "Indicates if the command is in b64",
                    "type": "boolean"
                },
                "description": {
                    "description": "The description from the workflow step",
                    "type": "string"
                }
            }
        },
        "api.ManualOutArgsUpdatePayload": {
            "type": "object",
            "required": [
                "response_out_args",
                "response_status",
                "type"
            ],
            "properties": {
                "response_out_args": {
                    "description": "Map of cacao variables storing the out args value, handled in the step out args, with current values and definitions",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cacao.Variables"
                        }
                    ]
                },
                "response_status": {
                    "description": "Indicates status of command",
                    "allOf": [
                        {
                            "$ref": "#/definitions/manual.ManualResponseStatus"
                        }
                    ]
                },
                "type": {
                    "description": "The type of this content",
                    "type": "string",
                    "example": "string"
                }
            }
        },
...
        "cacao.Addresses": {
            "type": "object",
            "additionalProperties": {
                "type": "array",
                "items": {
                    "type": "string"
                }
            }
        },
        "cacao.AgentTarget": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/cacao.Addresses"
                },
                "agent_target_extensions": {
                    "$ref": "#/definitions/cacao.Extensions"
                },
                "authentication_info": {
                    "type": "string"
                },
                "category": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "contact": {
                    "$ref": "#/definitions/cacao.Contact"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/cacao.CivicLocation"
                },
                "logical": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "sector": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "cacao.AuthenticationInformation": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kms": {
                    "type": "boolean"
                },
                "kms_key_identifier": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "oauth_header": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private_key": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "cacao.CivicLocation": {
            "type": "object",
            "properties": {
                "administrative_area": {
                    "type": "string"
                },
                "building_details": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "network_details": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "precision": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "street_address": {
                    "type": "string"
                }
            }
        },
...
        "cacao.Contact": {
            "type": "object",
            "properties": {
                "contact_details": {
                    "type": "string"
                },
                "email": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "phone": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "cacao.Extensions": {
            "type": "object",
            "additionalProperties": true
        },
...
        "cacao.Variable": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "constant": {
                    "description": "Indicate if it's a constant",
                    "type": "boolean",
                    "example": false
                },
                "description": {
                    "description": "A description of the variable",
                    "type": "string",
                    "example": "some string"
                },
                "external": {
                    "description": "Indicate if it's external",
                    "type": "boolean",
                    "example": false
                },
                "name": {
                    "description": "The name of the variable in the style __variable_name__ (not part of CACAO spec, but included in object for utility)",
                    "type": "string",
                    "example": "__example_string__"
                },
                "type": {
                    "description": "Type of the variable should be OASIS  variable-type-ov",
                    "type": "string",
                    "example": "string"
                },
                "value": {
                    "description": "The value of the that the variable will evaluate to",
                    "type": "string",
                    "example": "this is a value"
                }
            }
        },
        "cacao.Variables": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/cacao.Variable"
            }
        },
...
        "capability.ResolvedTarget": {
            "type": "object",
            "required": [
                "target"
            ],
            "properties": {
                "authentication": {
                    "$ref": "#/definitions/cacao.AuthenticationInformation"
                },
                "target": {
                    "$ref": "#/definitions/cacao.AgentTarget"
                }
            }
        },
...
        "manual.ManualResponseStatus": {
            "type": "string",
            "enum": [
                "success",
                "failure"
            ],
            "x-enum-varnames": [
                "ManualResponseSuccessStatus",
                "ManualResponseFailureStatus"
            ]
        },
...
    }
}